chore: improve CONTRIBUTING.md (#493)
* docs(contributing.md): added more useful commands to file

This commit aims to improve the commands avaible to the reader. New commands offer more ways of
consuming the repo while the old test commands have been replaced for commands that always produce
the cypress GUI (if you ignore the --skip-nx-cache flag, you only get the cypress GUI the first time
you run the command. Any time you run it afterwards you get a very meaningless message on the
terminal).

* docs(contributing.md): better readbility of sentence

Replaces "docs" for "documentation" as the file calls it that everywhere else.