FIX: added test files to tailwind config
This makes sure headless components behave the same on playwright as
they do in dev mode.

If removed, the checkbox component test that makes sure the icons are
hidden will fail as the appropiate tailwind classes won't be added *in
playwright mode only*. Unbeknownst to me, the tailwind classes are added
in dev mode.